Instead of arrows with variables and children, introduce the "each" node.
This changes templates of this form:

  authors as |author| -> <name type="personal">
    author.first -> <namePart type="given">

To this:

  each authors as |author|
    <name type="personal">
      author.first -> <namePart type="given">

Using the arrow when literals were not added to output was confusing. This change introduces a new node for the separate concept of creating a new context and evaluating children, but both still do iteration. This means that the arrow above can be written like this:

  each authors as |author|
    <name type="personal">
      each author.first as |first|
        <namePart type="given">
          first

But the arrow syntax is preferable because it is more succinct and avoids having to invent a new variable name.