Support requesting the client from paths starting with /forms.
This is so that we can give people URLs with paths starting with /forms rather than /forms2. Individual forms will be made available in production and rewrite rules added to make them available at paths starting with /forms. But we still want to be able to access forms at paths starting with /forms2 on the QA server.

- Prepend '/forms2/' to asset URLs. (Assets are still served under /forms2.)
- Don't use the HTML base element.
- Set the router's rootURL depending on the pathname used to request the client.